Rename addon-section.vue to align with other namings
