+++ bundles/org.openhab.ui/web/src/pages/settings/menu/addon-section.vue
